Move `logWallet`, add `logCommand`
