+++ packages/solidity-devops/js/utils/logger.js
@@ -1,18 +1,5 @@
 const chalk = require('chalk')
 
-const { getAccountBalance, getAccountNonce } = require('./chain.js')
-const { readWalletAddress, readWalletType } = require('./wallet.js')
-
-const logWallet = (chainName, walletName) => {
-  const walletAddr = readWalletAddress(walletName)
-  const walletType = readWalletType(walletName)
-  logInfo(`Wallet: ${walletAddr} [${walletName}, ${walletType}]`)
-  const balance = getAccountBalance(chainName, walletAddr)
-  const nonce = getAccountNonce(chainName, walletAddr)
-  logInfo(`  Balance: ${balance}`)
-  logInfo(`  Nonce: ${nonce}`)
-}
-
 const logSuccess = (msg) => {
   console.log(chalk.green.bold(msg))
 }
@@ -26,7 +13,11 @@ const logError = (msg) => {
 }
 
 const logInfo = (msg) => {
-  console.log(chalk.cyanBright(msg))
+  console.log(chalk.cyanBright.bold(msg))
 }
 
-module.exports = { logWallet, logSuccess, logWarning, logError, logInfo }
+const logCommand = (command) => {
+  console.log(chalk.blue(command))
+}
+
+module.exports = { logSuccess, logWarning, logError, logInfo, logCommand }
